Prevent init container failing when DB is empty
It's common for the init container to fail when cloning the repo
for the frst time or after deleting the postgres docker volume.
This is because it takes postgres a few seconds to initialise
the database and it isn't ready when the init container runs.

This ensure that the init container waits for the postgres container
to be ready, rather than simply running, before starting. Also added
to the web container.