Simplify Docker and Docker Compose files
For the time being we don't need two docker-compose files. Simplify
and re-order the Dockerfile (mostly to simplify the container's
internal directory structure).

Add a local-init.sh script which is not copied into the app image
but run in the init container via a bind mount. In addition to the
tasks previously carried out by the init container (migrate, seed),
this runs collectstatic. There's now no longer a need to do this
explicitly.

Remove references to deleted docker compose file in Makefile.

Update the README to reflect the above.