Remove the text "Warning" from warning
crispy-forms-gds has implementation of "Warning text" from the
govuk design system, but it is not up-to-date with it. The gov uk
design system doesn't show the text "Warning" in the Warning text,
just the warning text icon and the actual text.
This change makes the text "Warning" hidden to make it consistent
with govuk design system