Enable file download in admin site
The solution is necessarily complex because Django does not allow
easy manipulation of read-only fields, which is what we need for
reviewer users.

This generates download links on the fly which can be used to
retrieve the required file from storage and serve as a response.
This is required because while the application will have access
to S3, the browser will not.