Fix choice syntax in migration
The first tuple value should be machine readable (e.g. identifiers like
"central-gov") and the second human-visible "Central Government Department".

This is to make it clear which of the two is presented to the user and which is
for code only. That separation is also less confusing when it comes to content
design changes and translations.

See the official documentation for details:
https://docs.djangoproject.com/en/5.0/ref/models/fields/#choices