Send confirmation email
This commit is to send confirmation emails at Application
submitted page. This takes care of happy scenario.
The changes for email failure scenarios are to be added later.
It also reads NOTIFICATION_API_KEY from local .env file
Later on that is to be changed to pick it from AWS.

In order for the send email functionality to work, please
create a .env file at the project level and create two
following KEY value pair

NOTIFY_API_KEY=<provide api key>
SEND_EMAIL=True

The notification template is being developed and there will be
changes ( mostly on personalisation ) when the template is
finalised.

We need to be able to simulate sending email in the process
in the test scenario, e.g. when cypress test runs. In that case
we should not send the actual email. To cater for this, we are
using an email address :
"simulate-delivered@notifications.service.gov.uk" in the cypress
tests in the test cases where the application is submitted.

In order to use simulate-delivered@notifications.service.gov.uk,

we needed to add that email on the guest list here:

https://www.notifications.service.gov.uk/services/466c618f-6ded-4015-bea9-542931e2c659/api/guest-list

Please see https://docs.notifications.service.gov.uk/python.html#python-client-documentation

for more details on the Python Client API