Treat datetime fields as readonly
The built-in Django admin interface handles the auto-generated
time_submitted field properly. However the django-grappelli and
django-admin-interface packages, which were tested as quick ways
of making the admin site a bit nicer to work with, throw an
exception every time an application is loaded. This fixes that
by explicitly treating time_sumitted as readonly.

The time_decided field should be readonly for reviewer users,
so this is done too.