Add django-admin-interface theme
The django-admin-interface package is configurable via the UI with
the settings exportable like the contents of any other database
table.

This commit adds a json export of the settings in a place available
both in a running container and when running locally. The settings
are applied to the database on application start.