Add rejection and approval logic
Overriding the response_change method in the ModelAdmin subclass turns
out to be the best place to intervene in the admin flow to catch a
form submission and carry out some extra tasks. It ensures that any
changes to the model are saved as normal before the additional logic
is executed.

Because we're using a normal View class we need to decorate its
dispatch method to ensure it is subject to the admin site's
authentication. That it is also a subclass of ModelAdmin is a red
herring (it's normal View class methods to do its work) - this is
solely so we can use the message_user method.

The logic in the Review model which determines whether an application
is approvable/rejectable is effectively a placeholder, pending
discussions with the team. It's possible we want any application to
be rejectable.

Several new templates are required. The only notable one is a
confirmation page, triggered by rejecting or approving an application.