Get NOTIFY_API_KEY from environment variable
This will allow the NOTIFY_API_KEY to be provided from AWS, which
will be made available as the env variable NOTIFY_API_KEY

If the NOTIFY_API_KEY is available then it will send email,
else just log that NOTIFY_API_KEY is not found.

This is still taking care of the happy scenario only. Error
scenarios will be taken care in further PR.

Removed SEND_EMAIL flag as that is not needed.