Added S3Boto3Storage type for file uploads
Added the S3Boto3Storage type for file uploads on the AWS environment.
This is activated by setting the IS_AWS environment variable to true.

Added a log statement(s) to associate the session id with the file names, so we can identify which user initiated it.

Clamd is disabled on AWS. This should be reverted once we get the clamv instance deployed.