Set vars to make CSRF protection work with HTTPS
CSRF_COOKIE_SECURE makes sure that the cookie is marked as secure.  While
setting it to true normally ensures that the cookie will only be set when using
HTTPS, it can also happen that if not set, browsers won't send it over HTTPS
for security reasons.

CSRF_COOKIE_HTTPONLY ensures that the cookie isn't readable by JS, which
provides no actual protection but is sometimes required by security audits.

See the links above the definitions for further details