Add mypy to pre-commit checks
and make the necessary changes so mypy passes. There will be more type hints to
add as we go along.

Note that we set a dummy value to SECRET_KEY even if not running in DEBUG. This
is because pre-commit doesn't let you set DEBUG before running. This is not a
problem since we currently don't use SECRET_KEY, and even if we did an explicit
default value is better than the empty string.