Add model for Review guidance
The application review page will display guidance for reviewers. This
creates a model to store guidance for each review section, adds some
initial guidance in markdown format and populates the database with it
if no existing guidance is found.

When we create the new review form we will use the Python-Markdown
library to translate the markdown into HTML. This approach makes it
easy for superusers to edit the guidance.