Don't display the back-to-answers button on Registry details
Since Registry details is the last form in the journey, the Back To Answers
button did the same as the "Continue" button. Therefore we decided to remove
"Back To Answers".

So if a user completes the form but goes back to change the Registry information
they will only see a "continue" button that will take them back to Check your
answers.