Specify a Content Security Policy
In order to protect the application, a good thing to do is to tell the browser
not to load any asset (image, script or stylesheet) coming from any other origin
than the service itself (where the HTML comes from). This is done using
Content-Security-Policy, which adds an HTTP header to responses that tells the
browser what origins to accept.