Check that the key exists in the session dict
In some cases of visiting a URL the session has no registration_data, for
instance if the user enters a URL manually. Although this should be caught
upstream and return a 403, it's still better to check here if the key exist,
in order to avoid an unnecessary exception in the logs.