Keep user data in front-end, but discard at the end
Even when changing routes, some data should be kept. For instance,
the chosen domain name should be remembered even when the users changes
organisation type. Likewise uploaded files shouldn't be discarded when changing
answers to prevent the user having to upload them again on a different route.
However at the end of the user journey, when the data is about to be put in the
db, we need to remove anything left over that won't be needed and may be
confusing for reviewers.

This commit does 2 things:
1. don't remove anything the user's entered, even if they change routes.
2. removing anything not needed at the end when the application is submitted.