Add a label for accessibility
For the 3 upload questions we add a label before the form, in order to make
it easier for assistive tech users to link the question to the form controls.

Since the question in the label is already the title of the page above, we make
it invisible for sighted-navigation, to avoid repetition.