Introduce 'nonce' attribute to script elements.
Updated the code to use the facility provided by the csp module to use the generated nonce value in the inline script and style elements.

This was done so that in the admin screens we use javascript to handle record filtering and it was not possible to get it working due to csp settings. Therefore, we have modified/ overridden the templates to use nonce attribute in the style and script elements.