Fixed removal of minister session data
When user take a route where the answer for minister support is "No"
e.g. when the route is central gov->Email->Minister-No, we are
getting an error:
notifications_python_client.errors.HTTPError: 400 - [{'error': 'BadRequestError', 'message': 'Missing personalisation: minister'}]`

This is because the corresponding session data "minister" was being
deleted. We should not remove the session data there, as it is later
used during personalisation of corresponding email template, i.e.
we should preserve the answer "no" provided by the user.

This change removes deletion of the minister session data in that
scenario.