Validate the session data before sending to db
Depending on the route the user took, the session data fields
should have a specific value. And because sometimes the user
changes their answers, some fields are removed or added. Therefore
we need to make sure that the session data is correct or not before
we save it in the database and send it for further processing.