Check email failure
When email send to registrar email address is failed ( e.g. if
it is wrong/non-existent email address) then we need to update
the application status saying so and send an email to internal
team so that they are aware of the issue.

This has been implemented as django management command temporarily,
When celery is added to the project then the main method being
called i.e. "check_email_failure_and_notify" should be made
a celery shared task and should be run periodically.

I have added unique=True to Application model's reference field
as application is being retrieved using reference in the
email fail check process and this will create an index which will
help with faster retrieval.

Also refactored getting notification api client in a separate method
as that is being used during email fail check process