Fixed the appliation/review filtering
The issue was that we were not correctly using the custom filter - which lead to the lists not being populated. Removed the custom filter and used the Django template overriding to achieve the same without needing a custom class. Updated the javascript which attaches the event listener as it was not connecting to all elements too.