Reset the session when the users clicks the start button
We keep the session data throughout the user journey, whatever the path they
take, until they submit their application. This is so that they never have to
re-type the same thing twice even if they change routes.

However an exception to this should be when the user goes back to the start
page and clicks the Start button.

In order to do that, the Start button now goes to a different url. The view for
that URL resets the and then redirects to the first form (i.e. registrar
details).