Enable superadmins to hide registrars
When new registrars are onboarded, they can be added in the admin
site.

When they are offboarded we do not want to delete them as they may
be attached to old applications. This enables them to be marked
as inactive, which in turn hides them in the end-user dropdown.
It prevents them being deleted altogether.