Update the initial list of registrars
This removes all existing registrars since at least one was incorrectly
named. It adds the Nominet list as of 12/06/24.

Shortly after go-live the contents of the handle function should be
changed to the commented-out code. Until then, any changes to
registrars made in the admin site will not be preserved.