Celery to run periodic task
We need to call check_email_failure_and_notify every minute to
check for any failed email and send notification email on
top of that

This task is to integrate celery to do that.

At the moment these changes are the minimum needed to add celery.

We might have to make further changes based on learnings when we
try to deploy these.