Added env variables to docker-compose
Even though we don't need the nominet environment
variables to run the app locally, they need to be declared
for the application to run. Here we add them to docker-compose.yml
so we're able to continue using the app locally using docker.

without those two variables the approval process does not work under docker