Admin screen approval bug and tests
The issue/bug was that the review id was taken as the application id when marking an application as approved/rejeted.

Added Django test to check the application approval through the review screen. We create two applications through the usual client workflow as well as one directly from the admin screen.

Then try to approve the last applications created by the client. This test validates the approval screen contains the correct data.