Remove non-alphanumeric characters from uploaded file name.
There was a bug caused by uploading a file with the name containing non-alphanumeric characters in its name. Behind the scenes, it seems boto is trying to clean the names and this created mismatch between the name the system has and the actual name stored in the S3 bucket. Therefore, we are cleaning up the name so the name will be consistent.