Cookie Banner for analytics
This presents a cookie banner, similar to the one on gov.uk. Only if the
user accepts cookies (and the GOOGLE_ANALYTICS_ID is set) do we load the
Google Analytics snippet.

The user can go to the cookies page if they change their mind.
Because we don't know what cookies Google Tag Manager can set we
can't delete them if the user changes their mind and rejects them.