Add a bogus google analytics tracking ID for tests
The celery tests that check the cookie banner need the
google analytics snippet to be in the HTML in order to
make sure it's not executed when rejecting cookies.

Whether the snippet is included in the HTML is decided
by the existence of the GOOGLE_ANALYTICS_ID environment
variable.

Therefore in order to run the tests when the app is running locally in
docker , the container must have that variable set, with value which
won't actually send data to our GA account.