Custom information for review history
Customise the review history so that it shows the summary of application status changes along with the changes done in the review.

To provide this, we had to override the /simple_history/object_history.html so we keep the default behaviour for all other types.

In the status summary table, we filter out similar records and only display the latest record of the similar group. For detailed information, the user can view the relevant application's history under Applications.