Pin postgres major version to same as that in RDS
In the current code, the postgres version
is unpinned and hence while using the docker compose
commands, it uses the latest version of postgres which happens
to be v18 as of today.

It has been noticed that if we let it work on v18, then it needs
adjustment in volumes configuration, otherwise
docker compose up fails.

However, since the RDS postgres major version is 16, it is better
to pin it to 16 in docker-compose, so that local and RDS major
versions are same and there is a less chance of any issues due to
mismatch.