Upgrade Django to 4.2.27
This is to fix CVE-2025-64460 and CVE-2025-13372, both of which
will get resolved by django 4.2.27

These are listed in Security Tab of the repository, however for
some reason, dependabot hasn't created the PR for these, hence
creating the PR manually.

The details for the CVEs are:

CVE-2025-64460 - Django, django:
An issue was discovered in 5.2 before 5.2.9, 5.1 before 5.1.15,
and 4.2 before 4.2.27. Algorithmic complexity in
`django.core.serializers.xml_serializer.getInnerText()` allows
a remote attacker to cause a potential denial-of-service attack
triggering CPU and memory exhaustion via specially crafted XML
input processed by the XML `Deserializer`. Earlier, unsupported
Django series (such as 5.0.x, 4.1.x, and 3.2.x) were not evaluated
and may also be affected. Django would like to thank Seokchan
Yoon for reporting this issue.

CVE-2025-13372 - Django, django:
An issue was discovered in 5.2 before 5.2.9, 5.1 before 5.1.15,
and 4.2 before 4.2.27. `FilteredRelation` is subject to SQL
injection in column aliases, using a suitably crafted dictionary,
with dictionary expansion, as the `**kwargs` passed to
`QuerySet.annotate()` or `QuerySet.alias()` on PostgreSQL. Earlier,
unsupported Django series (such as 5.0.x, 4.1.x, and 3.2.x)
were not evaluated and may also be affected. Django would like
to thank Stackered for reporting this issue.